perf: Improve Benefits component
add theme's variables, add mobile view, map method perfomanced

PEIT-24 Agregar variables del theme a los componentes
