+++ src/Components/index.js
@@ -2,8 +2,8 @@ export { default as Ads } from './Ads/Ads'
 export { default as Banner } from './Banner/Banner'
 export { default as ImageSlider } from './Banner/ImageSlider'
 export * from './Banner/SlideData'
-export { default as Benefits } from './Benefits/Benefits'
 export { default as FollowUs } from './FollowUs/FollowUs'
+export { default as Benefits } from './Layout/Benefits/Benefits'
 export { default as AccordionFooter } from './Layout/Footer/AccordionFooter'
 export { default as Footer } from './Layout/Footer/Footer'
 export { default as Header } from './Layout/Header/Header'
